Rerender non-floating Lexical toolbar on selection change
Fixes LB-1597
